Create Constructors
Breakdown of the Code:
Default Constructor:

Defined as public Person().
Initializes Name to "Unknown" and Age to 0.
A message is printed to indicate that this constructor has been called.
Parameterized Constructor:

Defined as public Person(string name, int age).
Initializes Name and Age with the provided values.
A message is printed to indicate that this constructor has been called.
Static Constructor:

Defined as static Person().
It is called only once when the class is first used. It can only access static members and cannot have parameters.
A message is printed to indicate that the static constructor has been called.
Copy Constructor:

Defined as public Person(Person other).
Creates a new Person object using the values from an existing Person object.
A message is printed to indicate that this constructor has been called.
Main Method:

Demonstrates the use of different constructors by creating instances of the Person class and displaying their information.
This example provides a clear view of how constructors are used in C# to initialize objects in various ways.