Create Abstract Class
In C#, an abstract class is a class that cannot be instantiated and is meant to be inherited by other classes. Abstract classes can have both abstract methods (which must be implemented by derived classes) and regular methods (with implementations). The primary purpose of abstract classes is to define common behavior that can be shared by subclasses while allowing them to implement their specific behavior.
In this example, Animal is an abstract class with an abstract method Sound(). The subclasses Dog and Cat must provide their own implementations of the Sound() method. The Sleep() method in the abstract class has an implementation that is inherited by the derived classes.