Create C# code to implement MQTT publish and subscribe:
C# program that demonstrates how to use MQTT for publishing and subscribing to a topic. This example uses the popular MQTTnet library, which you can install via NuGet in your C# project.

1. Install the MQTTnet package from NuGet:



Install-Package MQTTnet

2. C# code to implement MQTT publish and subscribe:
Explanation:

Broker: We are connecting to a public MQTT broker (broker.hivemq.com) for demonstration.

Publish: The client publishes the message "Hello, MQTT!" to the topic "test/topic".

Subscribe: The client subscribes to the topic "test/topic" and prints any messages received.

Message Handling: When a message is received, the payload is decoded and displayed in the console.


You can run this code to experiment with MQTT, and the client will subscribe and publish messages to the specified topic.